2003-05-12  Willem Jan Palenstijn <wjpalenstijn@users.sourceforge.net> 	* world/Item.*: loopscript support in Item 	* world/*: changed some fields in Item to uint16 	* usecode/UCMachine.*: one (out of three) loopscript opcodes 	* usecode/UCStack.h: UCStack::push1(uint8)

+++ world/ItemFactory.cpp
@@ -30,7 +30,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include "Egg.h"
 
 Item* ItemFactory::createItem(uint32 shape, uint32 frame, uint16 quality,
-							  uint32 flags, uint32 npcnum, uint32 mapnum,
+							  uint16 flags, uint16 npcnum, uint16 mapnum,
 							  uint32 extendedflags)
 {
 	// check what class to create
@@ -115,7 +115,7 @@ Item* ItemFactory::createItem(uint32 shape, uint32 frame, uint16 quality,
 
 
 Actor* ItemFactory::createActor(uint32 shape, uint32 frame, uint16 quality,
-								uint32 flags, uint32 npcnum, uint32 mapnum,
+								uint16 flags, uint16 npcnum, uint16 mapnum,
 								uint32 extendedflags)
 {
 	// this function should probably differentiate between the Avatar,
